Final improvements to fix_bad_archives utility
`fix_bad_archives` now has a `force_update` flag that allows for forced
tagging of objects, in the case where they are already in the archival
storage class, but they do not yet have the `archived: true` tag applied.
It also now reports both storage class and tag status in the stats.