Fixes incorrect use of `get_tator_store`
Previously, it was possible for the call to `get_tator_store` to provide
a bucket and set the `upload` argument to `True`, which causes
`get_tator_store` to raise an exception. This changes the logic to
provide a project-specific upload bucket _or_ set the `upload` argument
to `True`, but not both.